separate package for scrapers
